+++ src/main/java/ml/pluto7073/bartending/foundations/tags/BartendingTags.java
@@ -17,6 +17,8 @@ public class BartendingTags {
     public static final TagKey<Item> BOILABLES = item("boilables");
     public static final TagKey<Item> BOTANICAL_ELEMENTS = item("botanical_elements");
     public static final TagKey<Item> C_FRUITS = item("c:fruits");
+    public static final TagKey<Item> WINE_FRUITS = item("bartending:wine_fruits");
+    public static final TagKey<Item> NETHER_WINE_FRUITS = item("bartending:nether_wine_fruits");
     public static final TagKey<Item> WARPED_BOTANICALS = item("warped_botanicals");
     public static final TagKey<Item> CRIMSON_BOTANICALS = item("crimson_botanicals");
     public static final TagKey<Item> GLASSES = item("glasses");